Refactor to better name, 'select'
