+++ AdminWebsite/AdminWebsite/ClientApp/src/app/shared/select/index.ts
@@ -1,0 +1,1 @@
+export * from './select.component';
