+++ AdminWebsite/AdminWebsite/ClientApp/src/app/work-allocation/allocate-hearings/allocate-hearings.component.html
@@ -30,7 +30,7 @@
       </div>
       <div class="govuk-grid-column-one-third">
         <app-justice-users-menu #csoFilterMenu (selectedEmitter)="onJusticeUserForFilterSelected($event)"></app-justice-users-menu>
-        <app-menu
+        <app-select
           #csoFilterMenu2
           (itemSelected)="onJusticeUserForFilterSelected2($event)"
           [items]="justiceUsersMenuItems"
@@ -38,19 +38,19 @@
           placeholder="Select users"
           title="Allocated CSO"
           [selectedEntityIds]="selectedJusticeUserIds"
-        ></app-menu>
+        ></app-select>
       </div>
     </div>
     <div class="govuk-grid-row govuk-!-margin-bottom-4">
       <div class="govuk-grid-column-one-third">
         <app-case-types-menu (selectedEmitter)="onCaseTypeSelected($event)"></app-case-types-menu>
-        <app-menu
+        <app-select
           (itemSelected)="onCaseTypeSelected2($event)"
           [items]="caseTypesMenuItems"
           ariaLabel="Case types list"
           placeholder="Select case types"
           title="Case types"
-        ></app-menu>
+        ></app-select>
       </div>
       <div class="govuk-grid-column-one-third">
         <label class="govuk-label" for="case-number-entry">Case number</label>
@@ -81,7 +81,7 @@
             (selectedEmitter)="onJusticeUserForAllocationSelected($event)"
           >
           </app-justice-users-menu>
-          <app-menu
+          <app-select
             #csoAllocatedMenu2
             (itemSelected)="onJusticeUserForAllocationSelected2($event)"
             [items]="justiceUsersMenuItems"
@@ -89,7 +89,7 @@
             ariaLabel="User list"
             placeholder="Select user"
             title="Allocate to"
-          ></app-menu>
+          ></app-select>
         </div>
 
         <div *ngIf="messageCanBeDisplayed()" class="govuk-body govuk-!-font-weight-bold govuk-grid-column-one-half govuk-!-padding-top-6">
